third_party/blink/web_tests/platform/generic/fast/dom/Window/event-non-enumerable-expected.txt

PASS "event" in window is true
PASS window.propertyIsEnumerable("event") is false
PASS successfullyParsed is true

TEST COMPLETE
